+++ src/hooks/useAuthor.ts
@@ -2,11 +2,11 @@ import { type NostrEvent, type NostrMetadata, NSchema as n } from '@nostrify/nos
 import { useNostr } from '@nostrify/react';
 import { useQuery } from '@tanstack/react-query';
 
-export function useAuthor(pubkey: string | undefined) {
+export function useAuthor(pubkey: string | undefined, relays?: string[]) {
   const { nostr } = useNostr();
 
   return useQuery<{ event?: NostrEvent; metadata?: NostrMetadata }>({
-    queryKey: ['author', pubkey ?? ''],
+    queryKey: ['author', pubkey ?? '', relays ?? []],
     queryFn: async ({ signal }) => {
       if (!pubkey) {
         return {};
@@ -14,7 +14,9 @@ export function useAuthor(pubkey: string | undefined) {
 
       const [event] = await nostr.query(
         [{ kinds: [0], authors: [pubkey!], limit: 1 }],
-        { signal: AbortSignal.any([signal, AbortSignal.timeout(1500)]) },
+        relays && relays.length > 0
+          ? { signal: AbortSignal.any([signal, AbortSignal.timeout(4000)]), relays }
+          : { signal: AbortSignal.any([signal, AbortSignal.timeout(1500)]) },
       );
 
       if (!event) {